requirements.txt
aws-cdk-lib==2.174.0
constructs>=10.0.0,<11.0.0
diagrams
graphviz
pytest
pytest-mock
